[REFACTOR] Modify Help component to be in one file
