--- src/client/react/stages/Help/HelpMenu.js
@@ -1,19 +1,0 @@
-import React from 'react';
-import { Menu, MenuItem } from '@blueprintjs/core';
-
-
-class HelpMenu extends React.Component {
-	render() {
-		return (
-			<Menu>
-				<MenuItem text='Test'/>
-				<MenuItem text='Test'/>
-				<MenuItem text='Test'/>
-				<MenuItem text='Test'/>
-			</Menu>
-		);
-	}
-}
-
-
-export default HelpMenu;
